Separates cpu helpers and interrupt logic out
