+++ ls8/interrupt.h
@@ -1,0 +1,6 @@
+#ifndef _INTERRUPT_H_
+#define _INTERRUPT_H_
+
+int handle_interrupt(struct cpu *cpu);
+
+#endif
